Fixing unit test in delete field functionality - TRUNK-1929
git-svn-id: http://svn.openmrs.org/openmrs/trunk@17826 5bac5841-c719-aa4e-b3fe-cce5062f897a

+++ web/src/main/java/org/openmrs/web/controller/form/FieldFormController.java
@@ -80,7 +80,7 @@ public class FieldFormController extends SimpleFormController {
 			Field field = (Field) obj;
 			field = setObjects(field, request);
 			
-			if (action.equals(Context.getMessageSourceService().getMessage("general.delete"))) {
+			if (action != null && action.equals(Context.getMessageSourceService().getMessage("general.delete"))) {
 				try {
 					Context.getFormService().purgeField(field);
 				} catch (DataIntegrityViolationException e) {
